backend/requirements.txt
fastapi
pydantic
uvicorn
motor
typing-extensions
python-jose
cryptography
python-dotenv
